# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hospital CRM & Patient Engagement Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: channel
          in: query
          schema:
            type: string
        - name: doctorId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Segments ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Appointment:
      type: object
      required: [id, leadId, doctorId, specialization, sourceChannel, status, datetime, createdAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        leadName:
          type: ["string", "null"]
        leadMobile:
          type: ["string", "null"]
        doctorId:
          type: integer
        doctorName:
          type: ["string", "null"]
        specialization:
          type: string
        sourceChannel:
          type: string
          enum: [app_booking, web_booking]
        status:
          type: string
          enum: [booked, confirmed, completed, cancelled]
        datetime:
          type: string
        createdAt:
          type: string
    AppointmentInput:
      type: object
      required: [leadId, doctorId, specialization, sourceChannel, datetime]
      properties:
        leadId:
          type: integer
        doctorId:
          type: integer
        specialization:
          type: string
        sourceChannel:
          type: string
          enum: [app_booking, web_booking]
        datetime:
          type: string
    AppointmentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [booked, confirmed, completed, cancelled]
        datetime:
          type: string

    # ── Segment ────────────────────────────────────────────────────────────────
...